adding no_kube_apiserver methods
