+++ pkg/status/status.go
@@ -3,8 +3,6 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2018 Datadog, Inc.
 
-// +build !kubeapiserver
-
 package status
 
 import (
@@ -103,6 +101,31 @@ func GetCheckStatus(c check.Check, cs *check.Stats) ([]byte, error) {
 	}
 
 	return []byte(st), nil
+}
+
+// GetDCAStatus grabs the status from expvar and puts it into a map
+func GetDCAStatus() (map[string]interface{}, error) {
+	stats := make(map[string]interface{})
+	stats, err := expvarStats(stats)
+	if err != nil {
+		log.Errorf("Error Getting ExpVar Stats: %v", err)
+	}
+	stats["config"] = getPartialConfig()
+
+	stats["version"] = version.AgentVersion
+	stats["pid"] = os.Getpid()
+	hostname, err := util.GetHostname()
+	if err != nil {
+		log.Errorf("Error grabbing hostname for status: %v", err)
+		stats["metadata"] = host.GetPayloadFromCache("unknown")
+	} else {
+		stats["metadata"] = host.GetPayloadFromCache(hostname)
+	}
+	now := time.Now()
+	stats["time"] = now.Format(timeFormat)
+	stats["leaderelection"] = getLeaderElectionDetails()
+
+	return stats, nil
 }
 
 // getPartialConfig returns config parameters of interest for the status page
