+++ pkg/status/status_no_apiserver.go
@@ -1,0 +1,17 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2018 Datadog, Inc.
+
+// +build !kubeapiserver
+
+package status
+
+import (
+	log "github.com/cihub/seelog"
+)
+
+func getLeaderElectionDetails() map[string]string {
+	log.Info("Not implemented")
+	return nil
+}
